# Conflicts:
#	src/com/asd/framework/Appointment/WaitingAppointment.java
#	src/com/asd/framework/Calendar/Calendar.java
#	src/com/asd/framework/Calendar/CalendarTest.java

+++ src/com/asd/framework/Reminder/Reminder.java
@@ -16,7 +16,7 @@ public class Reminder {
 		r = new ReminderSMSStrategy(smsgateway, username, password);
 	}
 	
-	public boolean send(Integer receiverId, String text) {
+	public boolean send(Long receiverId, String text) {
 		return r.send(receiverId, text);
 	}
